Refactor frontend/actions to user sbp and move to controller/actions
