+++ frontend/controller/actions/index.js
@@ -1,0 +1,3 @@
+export { default as contract } from './contract.js'
+export { default as group } from './group.js'
+export { default as user } from './user.js'
